kafka/weather-forecast/consumer/requirements.txt
kafka-python-ng==2.2.3
psycopg[binary]==3.1.18
python-dotenv==1.0.0